backend/requirements.txt
fastapi
uvicorn
sqlmodel
asyncpg
google-generativeai
requests
python-dotenv
pydantic-settings
greenlet
